pyproject.toml
[tool.poetry]
name = "brd-agent"
version = "0.1.0"
description = "BRD Agent - converts meeting transcripts to structured BRD documents via Azure AI Foundry"
authors = ["Your Name <you@company.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
python-multipart = "^0.0.12"
azure-ai-inference = "^1.0.0b9"
azure-identity = "^1.19.0"
python-docx = "^1.1.2"
pydantic = "^2.9.0"
pydantic-settings = "^2.6.0"
httpx = "^0.27.0"
aiofiles = "^24.1.0"
python-dotenv = "^1.0.1"
loguru = "^0.7.2"
tenacity = "^9.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
httpx = "^0.27.0"
ruff = "^0.7.0"
mypy = "^1.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
